refactored each extractor to own module
